CMSPlugin refactoring
Move legacy and optional stuff into traits
