+++ libraries/src/Plugin/AutoloadLanguageAware.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * Joomla! Content Management System
+ *
+ * @copyright  (C) 2022 Open Source Matters, Inc. <https://www.joomla.org>
+ * @license    GNU General Public License version 2 or later; see LICENSE.txt
+ */
+
+namespace Joomla\CMS\Plugin;
+
+interface AutoloadLanguageAware
+{
+}
